StreamScope Launcher
SignalDataCollector.py has been updated so that at the beginning of the script, it will check to see if it has Admin privilages, and if not, will launch a new window and request privilages. It will then open a new window and start the streamscope server on the localhost, and then will continue running.